splits up downscale.py to dataset.py (dataset) and ds.py (delta downscaling).
